Fold table manipulation tests into structure, drop third file
The delete/select-all/wrapper tests fit naturally alongside the other
add/remove operations in structure.test.js.
